:memo: Docs: region 엔티티 수정 및 detective-office 폴더에서 office 폴더로 이름 변경
