+++ src/app.controller.ts
@@ -5,8 +5,8 @@ import { AppService } from './app.service';
 export class AppController {
   constructor(private readonly appService: AppService) {}
 
-  @Get()
-  getHello(): string {
-    return this.appService.getHello();
+  @Get('health')
+  async checkHealth(): Promise<string> {
+    return this.appService.checkDatabaseConnection();
   }
 }
