+++ src/app.service.ts
@@ -1,8 +1,16 @@
 import { Injectable } from '@nestjs/common';
+import { DataSource } from 'typeorm';
 
 @Injectable()
 export class AppService {
-  getHello(): string {
-    return 'Hello World!';
+  constructor(private readonly dataSource: DataSource) {}
+
+  async checkDatabaseConnection(): Promise<string> {
+    try {
+      await this.dataSource.query('SELECT 1');
+      return 'Database connection is healthy';
+    } catch (error) {
+      return `Database connection error: ${error.message}`;
+    }
   }
 }
