--- src/detectiveoffice/detectiveoffice.controller.ts
@@ -1,7 +1,0 @@
-import { Controller } from '@nestjs/common';
-import { DetectiveofficeService } from './detectiveoffice.service';
-
-@Controller('detectiveoffice')
-export class DetectiveofficeController {
-  constructor(private readonly detectiveofficeService: DetectiveofficeService) {}
-}
