--- src/detectiveoffice/detectiveoffice.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { DetectiveofficeService } from './detectiveoffice.service';
-import { DetectiveofficeController } from './detectiveoffice.controller';
-
-@Module({
-  controllers: [DetectiveofficeController],
-  providers: [DetectiveofficeService],
-})
-export class DetectiveofficeModule {}
