--- src/detectiveoffice/detectiveoffice.service.ts
@@ -1,4 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class DetectiveofficeService {}
