+++ src/office/detectiveoffice.controller.ts
@@ -1,34 +1,7 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
-import { OfficeService } from './detectiveoffice.service';
-import { CreateOfficeDto } from './dto/create-office.dto';
-import { UpdateOfficeDto } from './dto/update-office.dto';
+import { Controller } from '@nestjs/common';
+import { DetectiveofficeService } from './detectiveoffice.service';
 
-@Controller('office')
-export class OfficeController {
-  constructor(private readonly officeService: OfficeService) {}
-
-  @Post()
-  create(@Body() createOfficeDto: CreateOfficeDto) {
-    return this.officeService.create(createOfficeDto);
-  }
-
-  @Get()
-  findAll() {
-    return this.officeService.findAll();
-  }
-
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.officeService.findOne(+id);
-  }
-
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateOfficeDto: UpdateOfficeDto) {
-    return this.officeService.update(+id, updateOfficeDto);
-  }
-
-  @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.officeService.remove(+id);
-  }
+@Controller('detectiveoffice')
+export class DetectiveofficeController {
+  constructor(private readonly detectiveofficeService: DetectiveofficeService) {}
 }
