+++ src/office/detectiveoffice.module.ts
@@ -1,14 +1,9 @@
 import { Module } from '@nestjs/common';
-import { OfficeService } from './detectiveoffice.service';
-import { OfficeController } from './detectiveoffice.controller';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Owner } from './entities/owner.entity';
-import { DetectiveOffice } from './entities/detective-office.entity';
-import { Location } from './entities/location.entity';
+import { DetectiveofficeService } from './detectiveoffice.service';
+import { DetectiveofficeController } from './detectiveoffice.controller';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Owner, DetectiveOffice, Location])],
-  controllers: [OfficeController],
-  providers: [OfficeService],
+  controllers: [DetectiveofficeController],
+  providers: [DetectiveofficeService],
 })
-export class OfficeModule {}
+export class DetectiveofficeModule {}
