--- src/office/dto/create-office.dto.ts
@@ -1,1 +1,0 @@
-export class CreateOfficeDto {}
