--- src/office/dto/update-office.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateOfficeDto } from './create-office.dto';
-
-export class UpdateOfficeDto extends PartialType(CreateOfficeDto) {}
