+++ src/office/entities/location.entity.ts
@@ -1,4 +1,4 @@
-import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
+import { Entity, PrimaryGeneratedColumn, Column, OneToMany } from 'typeorm';
 import { DetectiveOffice } from './detective-office.entity';
 
 @Entity({ name: 'location' })
@@ -6,10 +6,10 @@ export class Location {
   @PrimaryGeneratedColumn('increment', { type: 'bigint', unsigned: true })
   id: number;
 
-  @Column({ type: 'double precision', nullable: false })
+  @Column({ type: 'double precision', nullable: true })
   longitude: number;
 
-  @Column({ type: 'double precision', nullable: false })
+  @Column({ type: 'double precision', nullable: true })
   latitude: number;
 
   @Column({ type: 'varchar', length: 255, nullable: false })
